+++ scrapy/spiders/__init__.py
@@ -15,7 +15,7 @@ from scrapy.utils.trackref import object_ref
 from scrapy.utils.url import url_is_from_spider
 
 if TYPE_CHECKING:
-    from collections.abc import AsyncIterable, Iterable
+    from collections.abc import AsyncIterator, Iterable
 
     from twisted.internet.defer import Deferred
 
@@ -78,7 +78,7 @@ class Spider(object_ref):
         self.settings: BaseSettings = crawler.settings
         crawler.signals.connect(self.close, signals.spider_closed)
 
-    async def yield_seeds(self) -> AsyncIterable[Any]:
+    async def yield_seeds(self) -> AsyncIterator[Any]:
         """Yield the initial :class:`~scrapy.Request` objects to send.
 
         .. versionadded:: VERSION
@@ -126,6 +126,18 @@ class Spider(object_ref):
                 yield Request("https://b.example")
                 yield self.crawler.settings["SEEDING_POLICY"]
                 yield Request("https://c.example")
+
+        It is also possible to raise :exc:`~scrapy.exceptions.CloseSpider`, for
+        example to customize the close reason when there are no seeds to yield:
+
+        .. code-block:: python
+
+            async def yield_seeds(self):
+                seeds = await queue_service_client.get_seed_batch()
+                if not seeds:
+                    raise CloseSpider("no_seeds")
+                for seed in seeds:
+                    yield seed
 
         To write spiders that work on Scrapy versions lower than VERSION,
         define also a synchronous ``start_requests()`` method that returns an